backend/requirements.txt
Flask
requests
openai
python-dotenv 